+++ defmt/src/impls/core_/mod.rs
@@ -8,7 +8,6 @@
 mod alloc_;
 mod array;
 mod cell;
-#[cfg(feature = "ip_in_core")]
 mod net;
 mod num;
 mod ops;